GLSL refactor - some bugs still.
This entire commit was generate by ChatGPT Codex with the following 2 prompts:

PROMPT 1: Rewrite fuzzygraph.js to use GLSL for rendering fuzzy graphs. Keep the following function names, and continue to export them: parseEquationString, displayGraph, calcWindowBounds, getXWidth, makeLinearMapper. In index.html, make the fewest changes possible to make it work. Keep the vertex shader simple. Most of the work should be done as a fragment shader. Call this file fuzzygraph_glsl.js.

PROMPT 2: This seems to work, but is very slow - no speedup over the purely cpu-bound original version. The calculateFuncForWindow function is still using a CPU-bound 2-deep for-loop. This code should be moved to glsl also. Refactor things more if need be to move more of the computation (especially the most CPU-bound parts like those executed for every pixel) to GLSL code.